Tidying up warning messages (e.g. (void)ing unused return codes; dropping un-used variables); replacing an old-style boost bind with a lambda in NxsCharactersBlock::HandleTokenState.

+++ whtest/whtest_wrapper.cpp
@@ -51,7 +51,7 @@ int WHTest(Params &params, IQTree &tree) {
 
 	int retval;
 	size_t nseq = tree.aln->getNSeq();
-	int nsite = tree.aln->getNSite(); 
+    size_t nsite = tree.aln->getNSite(); 
 
 
 	WHT_setAlignmentSize(nseq, nsite);
